cr3bp planet as well as test suite improvment
