+++ pykep/trajopt/gym/tops/_tops_cr3bp.json
@@ -9,7 +9,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Halo 2 Halo in fixed tof",
         "tof_bounds": [5.0, 5.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P1": {
         "state_s": [1.1648780946517576, 0.0, -0.11145303634437023, 0.0, -0.20191923237095796, 0.0],
@@ -21,7 +24,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Halo to Halo",
         "tof_bounds": [2.0, 15.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P2": {
         "state_s": [1.1648780946517576, 0.0, -0.11145303634437023, 0.0, -0.20191923237095796, 0.0],
@@ -33,7 +39,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "???",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P3": {
         "state_s": [1.1648780946517576, 0.0, -0.11145303634437023, 0.0, -0.20191923237095796, 0.0],
@@ -45,7 +54,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Halo to DRO",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P4": {
         "state_s": [0.8289927126704472, 0.0, 0.0, 0.0, 0.49493397729663385, 0.0],
@@ -57,7 +69,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "tof_bounds": [2.0, 10.0],
         "info": "DRO to NHRO",
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P5": {
         "state_s": [1.05629995376,-0.000569644659014,-0.0977538595795,-0.000308616385147,0.197244892132,-0.00224408889733],
@@ -69,7 +84,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Transfer between bretzel orbits",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P6": {
         "state_s": [-0.49172053272512411537,-4.6548008635095167331e-06,2.1381198872668593103e-30,-1.7147323404433621481e-05,2.2621923262251604214,4.5166455445843393584e-30],
@@ -81,7 +99,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Earth-Moon resonant orbit to high inclination lunar orbit",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P7": {
         "state_s": [0.03694774113818046, 0.0, 0.0, 0.0, 6.16843737391734, 0.0],
@@ -93,7 +114,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "High amplitude planar Earth orbit to high inclination lunar orbit",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P8": {
         "state_s": [1.0744342525276502, 0.0, 0.14463388243408795, 0.0, -0.01557378854434334, 0.0],
@@ -105,7 +129,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Transfer between high inclination lunar orbits",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P9": {
         "state_s": [0.903608553094 , 0.00114198736952 , 0.132836987186 , 0.00137077475221 , -0.0297883487782 , 0.0221619999979],
@@ -117,7 +144,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Transfer between high inclination lunar orbits",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P10": {
         "state_s": [-1.6740479439595235, 0.28393321518884523, 0.0, 0.0, 1.357459865451631, 0.0],
@@ -129,7 +159,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "L4 vertical orbit to L4 planar orbits",
         "tof_bounds": [2.0, 15.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P11": {
         "state_s": [0.9930292301781378, 0.0, 0.04611015565352394, 0.0, 0.6768221764145949, 0.0],
@@ -141,7 +174,10 @@
         "mu_cr3bp": 0.01215058560962404,
         "info": "Halo to Halo",
         "tof_bounds": [2.0, 10.0],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P12": {
         "state_s": [0.586792825, 0.0, 0.0, 0.0, 0.956849854, 0.0],
@@ -153,7 +189,10 @@
         "mu_cr3bp": 0.0121506683,
         "info": "Oshima problem (doi: 10.1016/j.asr.2023.05.055) - DRO to DRO - fixed time",
         "tof_bounds": [12.4155005583399, 12.4155005583399],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     },
     "P13": {
         "state_s": [0.908282483, 0.0, 0.204570695, 0.0, -0.0552436507, 0.0],
@@ -165,6 +204,9 @@
         "mu_cr3bp": 0.0121506683,
         "info": "Oshima problem (doi: 10.1016/j.asr.2023.05.055) - Vertical to Lyapunov",
         "tof_bounds": [10.0, 20.698633859117347],
-        "m_s": 1.0
+        "m_s": 1.0,
+        "L": 384400000,
+        "TIME": 375000,
+        "MASS": 1000
     }
 }
