+++ pykep/udpla/CMakeLists.txt
@@ -1,2 +1,2 @@
-set(PYKEP_UDPLA_PYTHON_FILES __init__.py _tle.py _spice.py)
+set(PYKEP_UDPLA_PYTHON_FILES __init__.py _tle.py _spice.py _cr3bp.py)
 install(FILES ${PYKEP_UDPLA_PYTHON_FILES} DESTINATION ${_PYKEP_INSTALL_DIR}/udpla)
